requirements.txt
dash
datasets
flask
joblib
numpy
pillow
scikit-learn
spacy
torch
torchvision
tqdm
transformers
